Add secondary heap shrink
Mario Kart DS and Digimon World DS now boot

+++ retail/bootloader/source/arm7/main.arm7.c
@@ -612,7 +612,7 @@ int arm7_main(void) {
 	// File containing cached patch offsets
 	aFile patchOffsetCacheFile = getFileFromCluster(patchOffsetCacheFileCluster);
 	fileRead((char*)&patchOffsetCache, patchOffsetCacheFile, 0, sizeof(patchOffsetCacheContents));
-	u32 prevPatchOffsetCacheFileVersion = patchOffsetCache.ver;
+	u16 prevPatchOffsetCacheFileVersion = patchOffsetCache.ver;
 
 	int errorCode;
